Consolidate DB migrations
This commit consolidates the DB migrations into 1 migration per table.
This makes the tests run faster and the DB schema easier to read, but it
requires a `php artisan migrate:reset` before updating to this version
of the code, and a `php artisan migrate` after updating. Doing this will
delete all data in the DB, including all users and items.